+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -39,6 +39,7 @@ public protocol HIGTheme: Sendable {
     var pricingCard: any HIGPricingCardTokens { get }
     var chatBubble: any HIGChatBubbleTokens { get }
     var comment: any HIGCommentTokens { get }
+    var cover: any HIGCoverTokens { get }
     var dataTable: any HIGDataTableTokens { get }
     var breadcrumb: any HIGBreadcrumbTokens { get }
     var pageHeader: any HIGPageHeaderTokens { get }