+++ Sources/HIGThemesSystem/HIGAdminTheme.swift
@@ -46,6 +46,7 @@ public struct HIGAdminTheme: HIGTheme, Sendable {
     public let pricingCard: any HIGPricingCardTokens
     public let chatBubble: any HIGChatBubbleTokens
     public let comment: any HIGCommentTokens
+    public let cover: any HIGCoverTokens
     public let dataTable: any HIGDataTableTokens
     public let breadcrumb: any HIGBreadcrumbTokens
     public let pageHeader: any HIGPageHeaderTokens
@@ -175,6 +176,7 @@ public struct HIGAdminTheme: HIGTheme, Sendable {
         self.pricingCard = base.pricingCard
         self.chatBubble = base.chatBubble
         self.comment = base.comment
+        self.cover = base.cover
         self.dataTable = base.dataTable
         self.breadcrumb = HIGSystemBreadcrumbTokens(
             font: base.breadcrumb.font,